Move Deserializer to de module
